Fixed. Activityを一つに修正
